requirements.txt
fastapi
uvicorn
minio
pandas
numpy
scipy
pyyaml
cachetools
python-multipart